use grpc++ instead of grpc++_unsecure
With the recent grpc bump, grpc causes troubles when both grpc and grpc_unsecure are loaded by the same binary.
Here, it is the case when StfSender and StfBuilder load the OccLite plugin (which uses grpc).
This change fixes both and allows them to be controlled by ECS again, tested on a VM setup with FLP suite.

I do not know if it might break DataDistribution in some other context, I hope that anyone reviewing it may know...